+++ CHANGELOG.md
@@ -2,8 +2,31 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.0.0](https://github.com/paypal/paypal-messaging-components/compare/v0.4.1...v1.0.0) (2019-07-11)
+
+### Bug Fixes
+
+-   setup client id support ([#10](https://github.com/paypal/paypal-messaging-components/issues/10)) ([fca2dfb](https://github.com/paypal/paypal-messaging-components/commit/fca2dfb))
+-   custom validation refactor ([#11](https://github.com/paypal/paypal-messaging-components/issues/11)) ([b6f5fce](https://github.com/paypal/paypal-messaging-components/commit/b6f5fce))
+-   remove window.paypal dependency in legacy pipeline ([#8](https://github.com/paypal/paypal-messaging-components/issues/8)) ([d7ddd43](https://github.com/paypal/paypal-messaging-components/commit/d7ddd43))
+-   update legacy LE map ([#4](https://github.com/paypal/paypal-messaging-components/issues/4)) ([758c484](https://github.com/paypal/paypal-messaging-components/commit/758c484))
+
+### Features
+
+-   rerender on container attribute changes ([#9](https://github.com/paypal/paypal-messaging-components/issues/9)) ([cf2f794](https://github.com/paypal/paypal-messaging-components/commit/cf2f794))
+
+### [0.4.1](https://github.com/paypal/paypal-messaging-components/compare/v0.4.0...v0.4.1) (2019-06-19)
+
+### Bug Fixes
+
+-   use fetched custom markup to validate sign ([45dc002](https://github.com/paypal/paypal-messaging-components/commit/45dc002))
+
+## [0.4.0](https://github.com/paypal/paypal-messaging-components/compare/v0.3.7...v0.4.0) (2019-06-17)
+
+### Features
+
+-   SDK release ([3e4982c](https://github.com/paypal/paypal-messaging-components/commit/3e4982c))
+
 ### [0.3.7](https://github.com/paypal/paypal-messaging-components/compare/v0.3.6...v0.3.7) (2019-06-06)
-
-
 
 ### 0.3.6 (2019-06-04)